add processors directory, move attributes processor into separate file/module, add tests for spans, change signature for processors by type, i.e. process_logs, process_spans
